refactor(dev): simplify compiler (#4410)
* refactor(dev): rename getRouteModuleExportsCached to getRouteModuleExports

since caching is perf improvement that is purely an implementation detail.
also remove export for uncached route exports since that is not used
anywhere.

* refactor(dev): simplify build options

using string constant unions instead of enums

* refactor(dev): split up browser and server builds into their own modules

* refactor(dev): factor out logic for finding externals

* refactor(dev): compiler interfaces for browser and server

* refactor(dev): change server assets manifest plugin to accept an assets manifest promise

* refactor(dev): do not compile server incrementally

since it depends on the assets manifest

* refactor(dev): split up build and watch into separate modules

also, split out compile failure logging into its own module

* refactor(dev): implement `build` function based on new remix compiler

* refactor(dev): rename `BuildOptions` to `CompileOptions` and move into `compiler/` directory

* refactor(dev): watch

* refactor(dev): compiler index module

* refactor(dev): camelCase filenames

* refactor(dev): update imports and types for compiler in `commands`

* test(dev): fix type errors

* test(dev): remove outdated, unused test

test was hardcoded to be skipped everytime and referenced the old rollup-based compiler and old gist
app

* refactor(dev): rename `onBuildFailure` to `onCompileFailure`
